requirements-dev.txt
pytest
pytest-django
simplegmail
